default to pandas display when df.dtypes printed
